refactor: Turn ChatModel into ChatHandler
