+++ common/src/main/java/com/wynntils/handlers/chat/RecipientType.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.core.chat;
+package com.wynntils.handlers.chat;
 
 import java.util.regex.Pattern;
