+++ common/src/main/java/com/wynntils/handlers/chat/events/NpcDialogEvent.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.wynn.event;
+package com.wynntils.handlers.chat.events;
 
 import net.minecraft.network.chat.Component;
 import net.minecraftforge.eventbus.api.Cancelable;
