+++ common/src/main/java/com/wynntils/wynn/model/BombBellModel.java
@@ -6,7 +6,7 @@ package com.wynntils.wynn.model;
 
 import com.wynntils.core.managers.Model;
 import com.wynntils.mc.utils.ComponentUtils;
-import com.wynntils.wynn.event.ChatMessageReceivedEvent;
+import com.wynntils.handlers.chat.events.ChatMessageReceivedEvent;
 import com.wynntils.wynn.objects.BombInfo;
 import com.wynntils.wynn.objects.BombType;
 import java.util.Set;
